feat: update existing tests
