+++ package.json
@@ -50,7 +50,7 @@
   "devDependencies": {
     "@nestjs/cli": "^9.0.0",
     "@nestjs/schematics": "^9.0.0",
-    "@nestjs/testing": "^9.1.6",
+    "@nestjs/testing": "^9.4.0",
     "@types/bcrypt": "^5.0.0",
     "@types/bull": "^4.10.0",
     "@types/express": "^4.17.13",
